Path simplification/wire merging, created path_util.rs, moved simplifier and offsetting code there.
