+++ homotopy-graphics/src/lib.rs
@@ -46,6 +46,7 @@ pub mod buffers;
 pub mod geom;
 pub mod gl;
 pub mod parity;
+pub mod path_util;
 pub mod subdivision;
 pub mod svg;
 pub mod tikz;
